% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_stream.R
\docType{data}
\name{moment_stream}
\alias{moment_stream}
\title{Online sample mean and sample variance computation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
moment_stream
}
\description{
\emph{The river where you set your foot just now is gone – those waters
giving way to this, now this.} ---Heraclitus
}
\details{
A \code{moment_stream} object is an \code{\link[R6]{R6}} object with
reference semantics. Rather than calculating the mean and variance on a
complete data set, which is difficult when the data does not fit in memory
or when only parts of it are available, \code{moment_stream}s use Welford's
online algorithm to calculate these values during a single pass through the
data (which can then be discarded if no other summary statistics are needed).

A stream can be monitored by creating an object with
\code{stream_name = moment_stream$new()}, and the mean and variance can be
calculated as new data comes in using \code{stream_name$update(x)}. The
sample mean is stored in \code{stream_name$m} and the sample variance is
stored in \code{stream_name$v_hat}.
}
\examples{
library(Heraclitus)
set.seed(1)
x = rnorm(1E4)
moments = moment_stream$new(x)
moments$m     #sample mean of x
moments$v_hat #sample variance of x

x2 = rnorm(1E3)
moments$update(x2)
all.equal(moments$m, mean(c(x, x2)))
all.equal(moments$v_hat, mean(c(x, x2)))
}
\keyword{datasets}
